+++ src/test/java/com/dmz/airdnd/common/auth/AuthControllerTest.java
@@ -19,7 +19,7 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 import com.dmz.airdnd.fixture.TestUserFactory;
-import com.dmz.airdnd.user.dto.request.response.UserRequest;
+import com.dmz.airdnd.user.dto.request.UserRequest;
 
 @WebMvcTest(AuthController.class)
 class AuthControllerTest {
